+++ src/components/header/Header.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
+import './Header.css';
 
 export default class Header extends Component {
   render() {
