+++ src/components/pages/about-page/AboutPage.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-export default class FourOhFour extends Component {
+export default class AboutPage extends Component {
   static propTypes = {};
 
   render() {
-    return <div>You done found the wrong path!</div>;
+    return <div />;
   }
 }
