+++ src/components/pages/about-page/index.js
@@ -1,0 +1,1 @@
+export { default } from './AboutPage';
