+++ src/components/pages/account-page/AccountPage.js
@@ -1,17 +1,16 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import SignupForm from '../forms/SignupForm';
-import LoginForm from '../forms/LoginForm';
+import SignupForm from '../../signup-form';
+import LoginForm from '../../login-form';
 
-export default class Account extends Component {
+export default class AccountPage extends Component {
   static propTypes = {};
 
   render() {
     return (
       <>
         <LoginForm onSubmit={this.props.onLogin} />
-
         <SignupForm onSubmit={this.props.onRegister} />
       </>
     );
