+++ src/components/pages/fourohfour-page/index.js
@@ -1,0 +1,1 @@
+export { default } from './FourOhFourPage';
