+++ src/components/pages/home-page/HomePage.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-export default class Home extends Component {
+export default class HomePage extends Component {
   static propTypes = {};
 
   render() {
