+++ src/components/pages/snippet-page/index.js
@@ -1,0 +1,1 @@
+export { default } from './SnippetPage';
