+++ src/components/search-bar/SearchBar.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './SearchBar.css';
 
 export default class SearchBar extends React.Component {
   constructor(props) {
