+++ src/components/snippet-list/SnippetList.css
@@ -1,0 +1,6 @@
+#snippets {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
