+++ src/components/snippet-list/SnippetList.js
@@ -1,6 +1,7 @@
 // 3. render Snips using state
 import React from 'react';
-import Snip from './Snip';
+import Snip from '../snippet';
+import './SnippetList.css';
 
 export default function SnipList({ snippets }) {
   console.log('Snip List rendered');
