+++ src/components/snippet/Snippet.css
@@ -1,10 +1,3 @@
-#snippets {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
 .snip {
   width: 90%;
   max-width: 609px;
