+++ src/css/style.css
@@ -1,21 +1,27 @@
-@import url("https://fonts.googleapis.com/css?family=Montserrat&display=swap");
-
-@import "resets.css";
-@import "palette.css";
-@import "navigation.css";
-@import "snippets.css";
-@import "search.css";
+/* Used for General Styling! Should have limited styles in here... */
 
 :root {
   font-size: 16px;
 }
 
 html {
-  font-family: "Montserrat", sans-serif;
+  font-family: 'Montserrat', sans-serif;
 }
 
 body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+
   background-color: var(--snips-white);
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
 }
 
 .info-block {
